requirements.txt
faiss_cpu==1.11.0
sentence_transformers==4.1.0
torch==2.5.1+cu121
tqdm==4.67.1
transformers==4.52.3
